Move types.ts to types/shared.ts and same with config.ts
