+++ frontend/src/configs/shared.tsx
