+++ frontend/src/pages/requests/config.tsx
@@ -10,7 +10,7 @@ import {
   getAgGridDateColFilterConfigs,
 } from "../../utils/agGrid";
 import { DownloadOption } from "../../hooks/useDownload";
-import { BuildDownloadOptionsParamsBase } from "../../types";
+import { BuildDownloadOptionsParamsBase } from "../../types/shared";
 
 type BuildDownloadOptionsParams = BuildDownloadOptionsParamsBase & {
   // Put additional parameters here if needed
