+++ frontend/src/types/shared.ts
